routes/image_ai.py
from flask import Blueprint, request, jsonify
from PIL import Image
import os
import random

image_bp = Blueprint("image_ai", __name__)

UPLOAD_FOLDER = "uploads"
os.makedirs(UPLOAD_FOLDER, exist_ok=True)

DISEASE_DATABASE = {
    "fungal": {
        "disease_name": "Fungal Skin Infection (Ringworm)",
        "causes": "Caused by dermatophyte fungi, moisture, sweat.",
        "effects": "Itchy circular rash, redness, scaling skin.",
        "precautions": "Keep skin dry, avoid sharing towels.",
        "treatment": "Apply antifungal cream. Consult doctor if severe."
    },
    "eczema": {
        "disease_name": "Eczema",
        "causes": "Allergens, stress, dry skin.",
        "effects": "Dry itchy inflamed patches.",
        "precautions": "Use moisturizers, avoid harsh soaps.",
        "treatment": "Topical steroids, antihistamines."
    },
    "acne": {
        "disease_name": "Acne Vulgaris",
        "causes": "Clogged pores, hormones.",
        "effects": "Pimples, inflammation.",
        "precautions": "Clean face regularly.",
        "treatment": "Salicylic acid or dermatologist visit."
    }
}

@image_bp.route("/analyze-image", methods=["POST"])
def analyze_image():

    if "image" not in request.files:
        return jsonify({
            "disease_name": "No Image Detected",
            "causes": "Image not uploaded properly",
            "effects": "Cannot analyze",
            "precautions": "Upload a clear image",
            "treatment": "Retry upload",
            "confidence": 0
        })

    file = request.files["image"]
    filepath = os.path.join(UPLOAD_FOLDER, file.filename)
    file.save(filepath)

    disease = random.choice(list(DISEASE_DATABASE.values()))

    return jsonify({
        "disease_name": disease["disease_name"],
        "causes": disease["causes"],
        "effects": disease["effects"],
        "precautions": disease["precautions"],
        "treatment": disease["treatment"],
        "confidence": round(random.uniform(0.80, 0.95), 2)
    })
