requirements.txt

flask
flask-cors
pymongo
twilio
requests
pillow
pdfplumber
geopy
